Don't activate Emacs if it isn't running
If you run Emacs and its server in a terminal, activating Emacs might
launch your GUI Emacs program, which is unwanted. Instead, check to see
if "Emacs" is a running application and only then activate it.

If no Emacs is running anywhere, the whole operation fails silently,
which is a known issue(?) or... Feature?