Fix property bug
Properties in an Applescript application, as it turns out, persist
between executions of the application, even if they are altered by the
application. A property's value is only reset when the application is
recompiled.

This behavior was causing the "-s" flag to emacsclient to be prepended
to the server path each time the application was run.

This change simply prepends the flag to a new variable and uses the
value of the property without changing it.